Work around broken copy-to-clipboard upon searching in Safari 13.
* Users reported that upgrading to Safari 13 broke the copy to clipboard
upon submitting the search query. Clicking on the search results to
copy to clipboard still worked, however.
* Safari 13 seems to no longer allow copying to clipboard in the
callback of a "synchronous" ajax request.
* Work around this by initializing a global, reused invisible text area,
populating that text area with the search results when they arrive, and
kicking off a setInterval upon searching to do the copy.
* This approach fails in Firefox: "document.execCommand('cut'/'copy')
was denied because it was not called from inside a short running
user-generated event handler", so only do this approach for Safari.